+++ app/src/main/java/woowacourse/payments/ui/cards/SingleCardsSection.kt
@@ -9,7 +9,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.ui.components.PaymentCard
 import woowacourse.payments.ui.components.PaymentCreateCard
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.preview.paymentCardUiModelSample
